Correct BackStack Issues
