+++ app/src/main/AndroidManifest.xml
@@ -4,10 +4,9 @@
 
     <uses-permission android:name="android.permission.INTERNET" />
     <uses-permission android:name="android.permission.ACCESS_NETWORK_STATE" />
+    <uses-permission android:name="android.permission.CAMERA" />
 
     <uses-feature android:name="android.hardware.camera" />
-
-    <uses-permission android:name="android.permission.CAMERA" />
 
     <application
         android:allowBackup="true"
@@ -16,32 +15,38 @@
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
+
         <activity
             android:name=".activity.MainActivity"
             android:label="MALA - UFRN"
-            android:theme="@style/AppTheme.NoActionBar">
+            android:theme="@style/TransparentTheme">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
+
         <activity
-            android:name=".activity.PrincipalActivity"
+            android:name=".activity.HubActivity"
             android:label="@string/app_my_loans"
             android:screenOrientation="portrait"
             android:theme="@style/AppTheme.NoActionBar" />
+
         <activity
-            android:name=".activity.NovoEmprestimoActivity"
+            android:name=".activity.NewLoanActivity"
             android:label="@string/app_new_loan"
             android:screenOrientation="portrait" />
+
         <activity
             android:name=".activity.NewLoanDetailsActivity"
             android:label="@string/app_new_loan"
             android:screenOrientation="portrait" />
+
         <activity
-            android:name=".activity.EmprestimoDetalheActivity"
+            android:name=".activity.LoanDetailsActivity"
             android:label="@string/app_my_loans_history_detail"
             android:screenOrientation="portrait" />
+
         <activity android:name=".activity.LogonActivity">
             <intent-filter>
                 <action android:name="android.intent.action.VIEW" />
@@ -55,10 +60,11 @@
                     android:scheme="http" />
             </intent-filter>
         </activity>
+
         <activity
             android:name=".activity.LoginActivity"
-            android:theme="@style/AppTheme.Dark" >
-        </activity>
+            android:theme="@style/AppTheme.Dark" />
+
     </application>
 
 </manifest>
