+++ app/src/main/res/layout/activity_loan_detail.xml
@@ -6,7 +6,7 @@
     android:orientation="vertical"
     android:paddingLeft="10dp"
     android:paddingRight="10dp"
-    tools:context="br.ufrn.mala.activity.EmprestimoDetalheActivity">
+    tools:context="br.ufrn.mala.activity.LoanDetailsActivity">
 
     <TextView
         android:id="@+id/lblCod_Barra_detail_loan"
