+++ app/src/main/res/layout/app_bar_act_principal.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".activity.PrincipalActivity">
+    tools:context=".activity.HubActivity">
 
     <android.support.design.widget.AppBarLayout
         android:layout_width="match_parent"
