+++ app/src/main/res/layout/content_act_principal.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     app:layout_behavior="@string/appbar_scrolling_view_behavior"
-    tools:context=".activity.PrincipalActivity"
+    tools:context=".activity.HubActivity"
     tools:showIn="@layout/app_bar_act_principal">
 
     <FrameLayout
