+++ app/src/main/res/values/styles.xml
@@ -44,4 +44,16 @@
         <item name="windowNoTitle">true</item>
     </style>
 
+    <style name="TransparentTheme" parent="Theme.AppCompat">
+
+        <item name="android:background">@null</item>
+        <item name="background">@null</item>
+
+        <item name="android:windowBackground">@android:color/transparent</item>
+        <item name="android:colorBackgroundCacheHint">@null</item>
+        <item name="android:windowContentOverlay">@null</item>
+        <item name="android:windowIsTranslucent">true</item>
+        <item name="android:windowAnimationStyle">@null</item>
+    </style>
+
 </resources>
